+++ cardano-cli/src/Cardano/CLI/EraBased/Run.hs
@@ -11,7 +11,6 @@ where
 import           Cardano.Api
 
 import           Cardano.CLI.EraBased.Commands
-import           Cardano.CLI.EraBased.Run.Address
 import           Cardano.CLI.EraBased.Run.Genesis
 import           Cardano.CLI.EraBased.Run.Governance
 import           Cardano.CLI.EraBased.Run.Query
@@ -20,6 +19,7 @@ import           Cardano.CLI.EraBased.Run.StakePool
 import           Cardano.CLI.EraBased.Run.TextView
 import           Cardano.CLI.EraBased.Run.Transaction
 import           Cardano.CLI.Helpers (printEraDeprecationWarning)
+import           Cardano.CLI.Run.Address
 import           Cardano.CLI.Run.Key
 import           Cardano.CLI.Run.Node
 import           Cardano.CLI.Types.Errors.CmdError
